Add PcrProductSequence, inherits from TemporarySequence
* PcrProductSequence is not a ChildSequence as its sequence is not a sub
sequence of the “parent” sequence.
* Simplify `getPcrProductsFromSequence` to use the SequenceModel
attribute.
