+++ public/scripts/plugins/pcr/lib/product.js
@@ -1,29 +1,15 @@
-import {assertIsInstance} from '../../../common/lib/testing_utils';
-import ChildSequence from './child_sequence';
-import Primer from './primer';
+import TemporarySequence from '../../../sequence/models/temporary_sequence';
 
 
-class Product extends ChildSequence {
-  setup(options) {
-    super.setup(options);
-    if(!(this.primer instanceof Primer)) {
-      this.primer.parentSequence = this.parentSequence;
-      this.primer = new Primer(this.primer, options);
-    }
-  }
-
+class PcrProductSequence extends TemporarySequence {
   get requiredFields() {
     return super.requiredFields.concat([
-      'primer',
+      'forwardPrimer',
+      'reversePrimer',
+      'stickyEnds',
     ]);
-  }
-
-  validate() {
-    assertIsInstance(this.primer, Primer, 'primer');
-    super.validate();
-    this.primer.validate();
   }
 }
 
 
-export default Product;
+export default PcrProductSequence;
