+++ public/scripts/plugins/pcr/plugin.js
@@ -1,5 +1,8 @@
 import Gentle from 'gentle';
 import PCRView from './views/pcr_view';
+import PcrProductSequence from './lib/product';
+import PcrPrimer from '../pcr/lib/pcr_primer';
+
 
 Gentle.addPlugin('sequence-primary-view', {
   name: 'pcr',
@@ -27,3 +30,10 @@ Gentle.addPlugin('sequence-canvas-context-menu', {
   },
   visible: Gentle.featureFlag('pcr')
 });
+
+
+// TODO, write v0to1 PreProcessor
+
+
+PcrProductSequence.registerAssociation(PcrPrimer, 'forwardPrimer', false);
+PcrProductSequence.registerAssociation(PcrPrimer, 'reversePrimer', false);
