+++ public/scripts/plugins/sequencing_primers/lib/product.js
@@ -1,0 +1,29 @@
+import {assertIsInstance} from '../../../common/lib/testing_utils';
+import ChildSequence from '../../pcr/lib/child_sequence';
+import Primer from '../../pcr/lib/primer';
+
+
+class Product extends ChildSequence {
+  setup(options) {
+    super.setup(options);
+    if(!(this.primer instanceof Primer)) {
+      this.primer.parentSequence = this.parentSequence;
+      this.primer = new Primer(this.primer, options);
+    }
+  }
+
+  get requiredFields() {
+    return super.requiredFields.concat([
+      'primer',
+    ]);
+  }
+
+  validate() {
+    assertIsInstance(this.primer, Primer, 'primer');
+    super.validate();
+    this.primer.validate();
+  }
+}
+
+
+export default Product;
